Implement approver tool wrapping, apply to inspect functions, create utils.py with printing functions.
- Add wrap_approvers() function for multi-stage approval process
- Apply approvers to bash and python functions from inspect
- Introduce python_allowlist_approver for enhanced Python security
- Move approval-related printing functions to utils.py

Enhances tool execution safety and improves code organization.
