+++ examples/intervention_approvals/intervention.py
@@ -1,15 +1,13 @@
-from rich.console import Group, RenderableType
-from inspect_ai.model._chat_message import ChatMessageAssistant
-from rich.markdown import Markdown
+from utils import print_tool_output, print_no_tool_response_and_ask_for_next_action
 from rich.panel import Panel
 from rich.prompt import Prompt
-from inspect_ai.solver._task_state import sample_state
+from typing import Optional
 from inspect_ai import Task, task
 from inspect_ai.dataset import MemoryDataset, Sample
 from inspect_ai.model import (
     ChatMessageTool,
     ChatMessageUser,
-    ModelOutput,
+    ChatMessageAssistant,
     call_tools,
     get_model,
 )
@@ -20,144 +18,17 @@ from inspect_ai.solver import (
     solver,
     system_message,
 )
-from inspect_ai.tool import Tool, ToolCall, python, tool
+from inspect_ai.tool import Tool, bash, python
 from inspect_ai.util import input_screen
-from inspect_ai.util import sandbox
 from approvals import (
     get_approval,
     Approver,
-    allow_list_approver,
+    bash_allowlist_approver,
     human_approver,
+    python_allowlist_approver,
+    wrap_approvers
 )
-from typing import Optional, List
 
-
-@tool
-def bash(timeout: int | None = None, user: str | None = None, approvers: Optional[list[Approver]] = None) -> Tool:
-    """Bash shell command execution tool.
-
-    Execute bash shell commands using a sandbox environment (e.g. "docker").
-
-    Args:
-      timeout (int | None): Timeout (in seconds) for command.
-      user (str | None): User to execute commands as.
-      approvers (Optional[list[Approver]]): The list of approvers to use.
-
-    Returns:
-      String with command output (stdout) or command error (stderr).
-    """
-
-    async def execute(cmd: str) -> str:
-        """
-        Use this function to execute bash commands.
-
-        Args:
-          cmd (str): The bash command to execute.
-
-        Returns:
-          The output of the command.
-        """
-        function_name = bash.__name__
-
-        if approvers:
-            state = sample_state()
-            if state is None:
-                return "Error: No state found."
-            
-            tool_calls = get_tool_calls_from_state(state)
-            if tool_calls is None:
-                return "Error: No tool calls found in the current state."
-
-            # Find the corresponding tool call
-            tool_call = next((tc for tc in tool_calls if tc.function == function_name), None)
-            
-            if tool_call:
-                approved, reason = get_approval(approvers, tool_call, state)
-            else:
-                return f"Error: No {function_name} tool call found in the current state."
-            
-            if not approved:
-                return f"Command rejected by the approval system. Reason: {reason}"
-        
-        with input_screen() as console:
-            console.print(Panel.fit(
-                f"Executing command: {cmd}",
-                title=f"{function_name.capitalize()} Execution",
-                subtitle="Command Approved"
-            ))
-        
-        # execute the command
-        result = await sandbox().exec(
-            cmd=["bash", "-c", cmd], timeout=timeout, user=user
-        )
-        # return output (including stderr if any)
-        output = ""
-        if result.stderr:
-            output = f"{result.stderr}\n"
-        return f"{output}{result.stdout}"
-
-    return execute
-
-def get_tool_calls_from_state(state: TaskState) -> Optional[List[ToolCall]]:
-    """
-    Safely extract tool calls from the last message in the state.
-    """
-    if not state.messages:
-        return None
-    last_message = state.messages[-1]
-    if isinstance(last_message, ChatMessageAssistant):
-        return last_message.tool_calls
-    return None
-
-def format_tool_call_output(output: list[ChatMessageTool]) -> list[RenderableType]:
-    output_renderables: list[RenderableType] = []
-    for i, tool_output in enumerate(output):
-        panel_contents = []
-        if tool_output.content != "":
-            panel_contents.append(Panel.fit(str(tool_output.content), title="Contents"))
-        if tool_output.error:
-            error_message = (
-                tool_output.error.message
-                if tool_output.error.message
-                else "error executing command"
-            )
-            panel_contents.append(
-                Panel.fit(error_message, title=f"Error ({tool_output.error.type})")
-            )
-
-        output_renderables.append(
-            Panel.fit(
-                Group(*panel_contents, fit=True),
-                title=f"Tool Output #{i} (ID: {tool_output.tool_call_id})",
-            )
-        )
-
-    return output_renderables
-
-def print_tool_output(output: list[ChatMessageTool]) -> None:
-    with input_screen() as console:
-        console.print(
-            Panel.fit(
-                Group(*format_tool_call_output(output), fit=True), title="Tool Output"
-            )
-        )
-
-def print_no_tool_response_and_ask_for_next_action(output: ModelOutput) -> str:
-    with input_screen() as console:
-        console.print(
-            Panel.fit(
-                Markdown(str(output.message.content)),
-                title="Model Response",
-                subtitle="No tool calls were generated.",
-            )
-        )
-
-    return Prompt.ask(
-        """[🖥️] The agent returned no tool calls.
-Please type a message to send to the agent, type nothing and hit enter to force another generation, or type 'exit' to end the conversation.
-""",
-        default="",
-    )
 
 @solver
 def get_user_prompt() -> Solver:
@@ -193,7 +64,7 @@ def agent_loop(tools: list[Tool], approvers: Optional[list[Approver]] = None) ->
                                 tool_call_id=tool_call.id,
                                 content=f"Command rejected by the approval system. Reason: {reason}"
                             ))
-                            continue  # Skip to the next tool call
+                            continue
                     
                     # Execute the approved tool call
                     single_tool_output = await call_tools(
@@ -259,13 +130,28 @@ Please reflect on what actions to take before you take them."""
 @task
 def tool_call_intervention():
     """
-    This is the intervention mode where you set approvers on tool call level.
+    This is the intervention mode where you set approvers on tool call level. We set bash to only allow ls command and escalate to human approver. For python we allow math, random, print, len, range, int, float, str, list, dict, set and escalate to human approver.
     """
-    allow_list_approver_func = allow_list_approver(["ls"])
+    bash_allow_list_approver_func = bash_allowlist_approver(["ls"])
     human_approver_func = human_approver()
-    approvers = [allow_list_approver_func, human_approver_func]
     
-    tools = [bash(approvers=approvers), python()]
+    # Create a Python allowlist approver
+    python_approver_func = python_allowlist_approver(
+        allowed_modules=['math', 'random'],
+        allowed_functions=['print', 'len', 'range', 'int', 'float', 'str', 'list', 'dict', 'set'],
+        disallowed_builtins={'eval', 'exec', 'compile', '__import__', 'open', 'input', 'globals', 'locals'},
+        sensitive_modules={'os', 'sys', 'subprocess', 'socket', 'requests', 'multiprocessing'},
+        allow_system_state_modification=False
+    )
+    
+    bash_approvers = [bash_allow_list_approver_func, human_approver_func]
+    python_approvers = [python_approver_func, human_approver_func]
+    
+    tools = [
+        wrap_approvers(bash(), approvers=bash_approvers),
+        wrap_approvers(python(), approvers=python_approvers)
+    ]
+    
     return Task(
         dataset=MemoryDataset([Sample(input="Unused in interactive mode")]),
         plan=[
@@ -279,7 +165,7 @@ def tool_call_intervention():
 @task
 def agent_loop_intervention():
     """
-    This is the intervention mode where you set approvers on agent level.
+    This is the intervention mode where you set approvers on agent level. We set human approver for all tool calls.
     """
     human_approver_func = human_approver()
     approvers = [human_approver_func]
